tests/requirements.txt
Deprecated==1.2.4
PyGithub==1.43.3
PyJWT==1.7.1
wrapt==1.10.11
